refactor initialState from reducer
